+++ lib/puncture_bb_impl.cc
@@ -1,17 +1,18 @@
 /* -*- c++ -*- */
-/* 
+/*
  * Copyright 2017 Moritz Luca Schmid, Communications Engineering Lab (CEL) / Karlsruhe Institute of Technology (KIT).
- * 
+ * Copyright 2018 Communications Engineering Lab (CEL) / Karlsruhe Institute of Technology (KIT).
+ *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 3, or (at your option)
  * any later version.
- * 
+ *
  * This software is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this software; see the file COPYING.  If not, write to
  * the Free Software Foundation, Inc., 51 Franklin Street,
@@ -56,7 +57,7 @@ namespace gr {
       d_vlen_in = puncturing_vector.size();
       d_vlen_out = ones(puncturing_vector);
       set_output_multiple(d_vlen_out);
-      set_relative_rate(d_vlen_out / d_vlen_in);
+      set_relative_rate(static_cast<float>(d_vlen_out) / static_cast<float>(d_vlen_in));
     }
 
     /*
@@ -95,4 +96,3 @@ namespace gr {
 
   } /* namespace dab */
 } /* namespace gr */
-
